requirements.txt
django>=1.7.1,<1.8
python-mpd2>=0.5.4
